// Angular/src/app/model/vehicles.ts
export class csVehicles{
    VehicleID: number ;
    ClientID: number ;
    VehicleRegistration: string ;
    VehicleRefNumber: string ;
    VehicleRefDate: Date ;
    VehicleMake: string ;
    VehicleModel: string ;
    VehicleTankCapacity: string ;
    VehicleEffectiveDate: Date ;
    VehiclePurchaseDate: Date ;
    VehicleCancellationDate: Date ;
    TemporaryRegistration: string ;
    VehiclesVIN: string ;
    FuelTypeID: number ;
    VehicleType: string ;
    OwnerType: string ;
    OwnerID: number ;
    UserModified: string ;
    Lastmodified: Date ;
    RegistrationReason: string ;
    Mode:string;
}